Added code for docking.  A successful dock requires the closing speed to be between 0.2ms and 0.4ms.  Any other speed will fail to dock and be treated as a collision.  Successful docking also requires proper alignment of the LM to the CSM
For collisions less then 2ms, half of the closing velocity will
be imparted to the csm and the lm will bounce off at half its
original closing velocity.  all bounces for now just reverse
the relative trajectory to the csm

For collsions greater than 2ms both vehicles are destroyed and
the simulation ends.