Removed dependancy on external files.  lunarref.txt is now included at compile time from lunarref.h.  userref.txt is still read at runtime, so user objects can still be placed there.
Default panels for when the .pnl files are not found are now the
new advanced panels and no longer the ams panel.  This allows for
the simulator to be run without any of the .pnl files and use the
advanced panels.  .pnl files can still be used to override the
defaults, so user panels are still possible.

This change primarly was to allow for the simulator to be run without
any additional files required.