Restructured source to accomodate new focus on project.  I have decided that LMS will remain purely a lunar mission simulator in the spirit of AMS and not have the Earth side of lunar missions.
The subproject Launch has been renamed to FMS for Full Mission
Simulator.  This will be an extended version of LMS which
includes Earth-based missions as well as lunar missions, possibly
in the future including the whole solar system which could allow
for missions to the other planetes either manned or probes.
Initially FMS will allow for simulation of missions starting
with Mercury/Redstone up through the Apollo lunar landings.
The Lunar portion of these missions will be identical to what
LMS simulates

Since so many modules are the same between the two different
simulators, a new common folder has been created to hold all
the objects that are common between both simulations.  This
prevents duplication of effort as well as bugs that can creep
in when maintaining two different code bases.